Finish section 25 - Advanced REST APIs

+++ backend/controllers/feed.js
@@ -4,6 +4,7 @@ const path = require("path");
 const { validationResult } = require("express-validator");
 
 const Post = require("../models/post");
+const User = require("../models/user");
 
 exports.getPosts = (req, res, next) => {
   const currentPage = req.query.page || 1;
@@ -48,21 +49,30 @@ exports.createPost = (req, res, next) => {
   const title = req.body.title;
   const content = req.body.content;
   const imageUrl = req.file.path.replace(/\\/g, "/");
+  let creator;
   const post = new Post({
     title,
     content,
     imageUrl,
-    creator: {
-      name: "John Doe",
-    },
+    creator: req.userId,
   });
 
   post
     .save()
-    .then((post) => {
+    .then(() => {
+      return User.findById(req.userId);
+    })
+    .then((user) => {
+      creator = user;
+
+      user.posts.push(post);
+      return user.save();
+    })
+    .then(() => {
       res.status(201).json({
         post,
         message: "Post created successfully!",
+        creator: { _id: creator._id, name: creator.name },
       });
     })
     .catch((err) => {
@@ -123,6 +133,11 @@ exports.updatePost = (req, res, next) => {
         error.statusCode = 404;
         throw error;
       }
+      if (post.creator.toString() !== req.userId) {
+        const error = new Error("Not authorized!");
+        error.statusCode = 403;
+        throw error;
+      }
 
       if (imageUrl !== post.imageUrl) {
         clearImage(post.imageUrl);
@@ -155,10 +170,23 @@ exports.deletePost = (req, res, next) => {
         error.statusCode = 404;
         throw error;
       }
+      if (post.creator.toString() !== req.userId) {
+        const error = new Error("Not authorized!");
+        error.statusCode = 403;
+        throw error;
+      }
+
       clearImage(post.imageUrl);
       return Post.findByIdAndDelete(postId);
     })
-    .then((result) => {
+    .then(() => {
+      return User.findById(req.userId);
+    })
+    .then((user) => {
+      user.posts.pull(postId);
+      return user.save();
+    })
+    .then(() => {
       res.status(200).json({
         message: "Post deleted successfully!",
       });
